+++ sparql_service/src/service_description_format.rs
@@ -1,0 +1,5 @@
+pub enum ServiceDescriptionFormat {
+    // Internal representation
+    Internal,
+    // TODO: add JSON, RDF, etc?
+}